pyproject.toml
[project]
name = "common-crawler"
version = "1.0.1"
description = ""
readme = "README.md"
authors = [{ name = "Minh Dao", email = "daoducminh1997@gmail.com" }]
requires-python = ">=3.12"
dependencies = [
    "httpx>=0.28.1",
    "itemloaders>=1.3.2",
    "pendulum>=3.1.0",
    "psycopg2-binary>=2.9.10",
    "pymongo>=4.16.0",
    "python-dotenv>=1.1.0",
    "requests>=2.33.1",
    "scrapy>=2.13.1",
    "sqlalchemy>=2.0.41",
    "sqlalchemy-cockroachdb>=2.0.2",
    "urllib3>=2.4.0",
]
